+++ CHANGELOG.md
@@ -7,7 +7,9 @@
 * `list`-command:
   * Fix displaying all files if filter does not match anything.
 * `push`-command:
-  * When specifying `--expire`, the notification about when the link expires is now printed to stderr instead of stdout.
+  * When specifying `--expire`, the notification about when the link expires is now
+    * printed to stderr instead of stdout.
+    * only printed if `asfa` is used in a terminal.
     This makes it easier to capture the URL in scripts like `tmp_link=$(asfa push my-file)`.
 
 ## v0.7.0 (2021-01-25)